requirements.txt
aiohttp==3.9.3
pydantic==2.6.1
pytest==8.0.0
pre-commit==3.6.0
scrapy==2.11.0
pymysql==1.1.0
python-dotenv==1.0.1 